requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
psycopg2-binary==2.9.9
apscheduler==3.10.4
slack-sdk==3.27.2
aiohttp==3.9.5
pydantic==2.7.1
